feat: RPX (xHash12) hash function implementation
